+++ src/IHM/model/ImperialSysLength.java
@@ -25,11 +25,11 @@ public class ImperialSysLength extends NonSI
 		}
 
 	}
-		public void fromReference(long qtOfRef)
+		public void fromReference(double qtOfRef)
 		{
 		}
 
-		public long toReference()
+		public double toReference()
 		{
 				return 0;
 		}
